Fix KOM/CR segment efforts not populating
Strava changed the API result data so I had to fix that. Then I made it
so segment efforts, if they exist already in the db, I'll update the
effort if a KOM or PR rank is set. That way if rankings changed
(refreshed activity achievement) it'll still get updated.