Fix issue with treadmill chart not showing correct data
Treadmill data was just being put into the first few charts shown since
there isn't treadmill data for every month. This was because of an issue
where I would run an executeQuery() on SQL statements, but not fetchAll.
I fixed this issue throughout the codebase.

The problem in this case stemmed from getRunningXAxisLabels() getting
called twice and that function tries to loop over the PDOStatement
returned from executeQuery. But the 2nd time you try and loop over it,
it doesn't work. By adding a fetchAll() after executeQuery() we can loop
over the results as much as needed since it's an array format instead of
PDOStatement.